Skip padsv op in $lex =~ ...
m//, s/// and y/// already have logic to deal with implicit lexical
$_.  The pad offset of $_ is stored in the match op itself.  We can
take advantage of that and extend it to lexical variables in general.
That way we have fewer ops to execute, as $lex =~ // no longer calls
pp_padsv.  It also allows lexical variables’ names to be mentioned in
uninitialized warnings for y///.