Call STORE on lexical $tied = vec/chr
The assignment gets optimised away, so pp_vec and pp_chr need to call
SvSETMAGIC on their targets.

This bug started happening with vec just recently, in db098081, when
the OPpTARGET_MY optimisation was applied to it.

The bug started happening with chr for the same reason, but back in
5.6.0 (b162f9ea).

I moved the STORE tests down in lex_assign.t when expanding them, to
avoid sabotaging the tests that check the results of the expressions
after the __END__ marker.