Don’t call STORE twice on setpgrp target
If the target is a lexical variable (as happens with ‘$lex = setpgrp’,
in which the assignment is optimised away), then doing set-magic mul-
tiple times has an observable effect.  This was only happening when it
was called with no arguments.

This double STORE goes back to 1f200948c4c.  1f200948c4c was just
meant to fix a stack bug, but extending the stack when necessary, but
it was also push TARG and setting it to -1, only for it to be over-
written shortly.  (Also, it didn’t fully fix the stack bugs.  See
88d6953212e.)