Don’t make temp copy for ()=...
List assignment will make temporary copies if either side contains ops
that *might* occur elsewhere on the stack.  That is obviously not nec-
essary for ()=....

()= is often used to count the number of items, suppress void warn-
ings, or force list context, so it’s worth optimising it.

It’s actually possible to test this without examining the op tree.
Just see whether we have redundant FETCH calls on items not assigned.