current_sub.t: Test sub(){__SUB__}
If the constant inlining algorithm changes, we need to make sure not
to break sub(){__SUB__} accidentally.

Sometimes __SUB__ gets optimised to a constant.  Constant inlining
searches the op tree for constants.  Currently we are saved by the
fact that the search happens before the peephole optimizer does the
conversion.