Deparse lexical subs
We currently have a problem with the hints to allow lexical subs to
begin with not necessarily being in scope at the point of the declara-
tion.  So the fix for that (which is a bit of a kludge) is to emit
‘use feature 'lexical_subs'’ and the equivalent of turning off the
corresponding warnings category, which may result in a surfeit of
^WARNING_BITS gibberish.  But at least it works. :-)

Also, package subs are not yet disambiguated with package name pre-
fixes if they fall in the same scope as lexical subs of the same name.