Remove vmsish pragma from one-liners in Test::Simple's exit.t.
Awaiting upstream application at:

https://github.com/Test-More/test-more/pull/494

This was broken by a recent change that added double quotes around
the double quotes that were already there.

It also hasn't been working as intended for these tests for some
time as they manipulate the exit code directly in an END block,
and that sneaks by the effects of the pragma when run at compile
time.

This works:

  $ perl  -e "require vmsish; import vmsish 'hushed'; END {$? = 1};"

This (which is what the tests are doing) doesn't work:

  $ perl -Mvmsish=hushed -e "END {$? = 1};"
  %SYSTEM-F-ABORT, abort

So we'll just live with the noise from the exit messages.  They
don't seem to cause any trouble for the core test suite.