Silence spurious locale warnings
When testing the locales on the system, we don't want to be warned that
particular ones may not work well with Perl, as the point of the test is
to find and list the incompatibilities.

When looking for locales suitable for some purpose, we try every one
possible until we find one that matches what we're looking for.  This
may entail briefly switching into locales that aren't fully compatible
with Perl.  Again we don't want to be warned about such a locale.  We're
not using it for anything except to determine suitability anyway.