Remove redundant EBCDIC lines from Deparse.t
When these were added to b.t in bd145f00387, the Deparse output still
contained \s, so it made sense to replace something else with \s in
order for the test to match.

In fcd95d642 and fee7e83848 it changed, such that the output no longer
contained \s, so, at least here, the EBCDIC code is doing nothing.