Escape EFS for directory portion in tovmsspec.
For some reason extended characters were only being escaped in
the final component of the path, but not in the directory portion,
the one exception being dots.  We need to give the entire path
the same treatment.  There is probably considerable opportunity
for further consolidation and refactoring in what we escape, but
at least now ../foo  bar/ correctly becomes [-.foo^_^_bar].

The reverse case has similar problems and is a TODO.