toke.c: Consistently upgrade under encoding
The documentation says that intermixing above-Latin1 code points with
ones that would be otherwise encoded to something else, like Greek,
causes the encoding to be foregone.  Until this commit, this only
happened when the above-latin1 code point came first in the string
constant being scanned; meaning string-order was important.  This
changes things to match the documentation