Remove extraneous semicolons from Deparse output
Prior to the previous commit, it was emitting extra semicolons for
statements optimised away outside of any subroutine, and extra semi-
colons after sub declarations and the end of the enclosing scope.  The
previous commit made it also emit an extra semicolon at the end of the
main program, if the last thing was a sub declaration.  So I decided
to fix all three at once.