Make B use B::COP for nulled COPs
Former COPs still carry information with them, and to get to it I had
to have B::Deparse rebless the op object into B::COP.

This commit makes B use the right class to begin with.  This happens
to make B::Concise output the extra information that nulled COPs carry
around with them, which was on my to-do list anyway, so I’m happy
about that.