[perl #77860] \& proto should disallow sub calls
It was checking to see whether its argument were an entersub op (&foo
is indeed an entersub) and then wrapping it in refgen (\).  But enter-
sub also covers &foo() and foo() and foo.  op_lvalue checks whether
the OPf_STACKED flag is set and only turns it into rv2cv if it is not.
(Only &foo lacks that flag.)  So I copied that logic into prototype
application.