reduce stderr noise in build
Ideally you should be able to do

    (Configure && make test) 2> /tmp/err

with /tmp/err being empty. This is not the case, and this commit
is a first step towards that goal.

The

    fprintf (stderr, "Sizeof time_t = %ld\n", sizeof (time_t));

in Configure appears to be just a debugging aid; it's also a copy of the
code in Porting/timecheck.c, which people can always run if need-be.

hints/linux.sh appears to be looking for symbols in libdb.so; if the
library is stripped, this produces to stderr:

    /bin/nm: /lib/libdb.so: no symbols

I've silenced it for now with a 2>/dev/null, but I'm not sure if the whole
test is flawed, since with no symbols, 'nm libdb.so | grep pthread' is a
no-op.

make_patchnum.pl, when running backticks, prints $? to stderr if it's
non-zero; since a similar print in the other branch is already commented
out, I assume its just left-over debugging.