Deparse $x =~ (1?/$a/:0) under taint mode
This code deparses incorrectly under taint mode:

$ ./perl -Ilib -mO=Deparse  -e '$x =~ (1?/$a/:0)'
$x =~ ($_ =~ /$a/);
-e syntax OK
$ ./perl -Ilib -mO=Deparse -T -e '$x =~ (1?/$a/:0)'
$x =~ /$a/;
-e syntax OK

The branch folding makes it deparse as ‘$x =~ /$a/’, whereas the /$a/
on the rhs, since it was not the argument to =~, is bound to $_, not
to $x.  That’s why B::Deparse adds the $_ =~, but it fails to do so
under taint mode.

It was broken by:

commit 7fb31b92fa6bf56dff7d4240b7051b9158f7df43
Author: David Mitchell <davem@iabyn.com>
Date:   Sun Apr 1 10:21:22 2012 +0100

    make OP_REGCRESET only for taint handling

    The OP_REGCRESET op, which is sometimes prepended to the chain of ops
    leading to OP_REGCOMP, currently serves two purposes; first to reset the
    taint flag, and second to initialise PL_reginterp_cnt. The second purpose
    is no longer needed, and the first has a bug, in that the op isn't
    prepended when "use re 'eval'" is in scope.

    Fix this by prepending the op solely when PL_tainting is in effect.
    This also makes run-time regexes slightly more efficient in the
    non-tainting case.

which has a typo in it.