[perl #123357] Fix deparsing of $; at stmt end
Instead of sometimes appending ; to statements and then removing it
later, to avoid doubling it up, *only* append ; to a statement when
actually joining them together or emitting them.  That fixes bugs with
do{$;} becoming do {$} and ‘$_=$;; $;=7;’ becoming ‘$_=$; $;=7;’.

I also removed the boilerplate from pp_stub, since it was redundant
(and slow) and also partially wrong.  The $name var was bogus.